Refactor pastas e arquivos
