--- .env
@@ -1,1 +1,0 @@
-PASSWORD_SUPER_SECRET = "123"
