+++ .gitignore
@@ -1,1 +1,3 @@
 node_modules
+.env
+config.json
