+++ package.json
@@ -18,8 +18,10 @@
     "nodemon": "^2.0.19",
     "path": "^0.12.7",
     "sequelize": "^6.21.6",
-    "sequelize-cli": "^6.4.1",
     "swagger-ui-express": "^4.5.0",
     "yup": "^0.32.11"
+  },
+  "devDependencies": {
+    "sequelize-cli": "^6.4.1"
   }
 }
