--- src/app/model/UserModel.js
@@ -1,23 +1,0 @@
-// import mongoose from "mongoose";
-const mongoose = require('mongoose');
-
-const UserSchema = new mongoose.Schema({
-    id: {
-        type: String,
-    },
-    nome: {
-        type: String,
-        required: true,
-    },
-    email: {
-        type: String,
-        required: true,
-    },
-    descricao: {
-        type: String,
-        required: true,
-    }
-});
-
-const UserModel = mongoose.model("users", UserSchema);
-module.exports =  UserModel;
