+++ src/app/routes/adminRoutes.js
@@ -1,5 +1,5 @@
 const { Router } = require('express');
-const AdminController = require('../controller/AdminController');
+const AdminController = require('../controllers/AdminController');
 
 const router = Router();
